+++ library/ssl_tls13_generic.c
@@ -883,222 +883,6 @@ static void mbedtls_ssl_create_verify_structure(
     *verify_buffer_len = 64 + content_string_len + 1 + transcript_hash_len;
 }
 
-
-/* mbedtls_ssl_tls1_3_derive_master_secret( )
- *
- * Generates the keys based on the TLS 1.3 key hierachy:
- *
- *                    0
- *                    |
- *                    v
- *     PSK ->  HKDF-Extract = Early Secret
- *                    |
- *                    v
- *     Derive-Secret( ., "derived", "" )
- *                    |
- *                    v
- *  (EC)DHE -> HKDF-Extract = Handshake Secret
- *                    |
- *                    v
- *     Derive-Secret( ., "derived", "" )
- *                    |
- *                    v
- *     0 -> HKDF-Extract = Master Secret
- *
- */
-int mbedtls_ssl_tls1_3_derive_master_secret( mbedtls_ssl_context *ssl )
-{
-    unsigned char ECDHE[66];
-
-    size_t ECDHE_len;
-    int ret = 0;
-    const mbedtls_md_info_t *md;
-    const mbedtls_ssl_ciphersuite_t *suite_info;
-    unsigned char *psk;
-    size_t psk_len;
-    unsigned int psk_allocated = 0;
-    int hash_size;
-
-    if( ssl->session_negotiate == NULL )
-    {
-        MBEDTLS_SSL_DEBUG_MSG( 1, ( "session_negotiate == NULL, mbedtls_ssl_tls1_3_derive_master_secret failed" ) );
-        return( MBEDTLS_ERR_SSL_INTERNAL_ERROR );
-    }
-
-    md = mbedtls_md_info_from_type( ssl->handshake->ciphersuite_info->mac );
-    if( md == NULL )
-    {
-        MBEDTLS_SSL_DEBUG_MSG( 1, ( "md == NULL, mbedtls_ssl_tls1_3_derive_master_secret failed" ) );
-        return( MBEDTLS_ERR_SSL_INTERNAL_ERROR );
-    }
-
-    suite_info = mbedtls_ssl_ciphersuite_from_id( ssl->session_negotiate->ciphersuite );
-    if( suite_info == NULL )
-    {
-        MBEDTLS_SSL_DEBUG_MSG( 1, ( "suite_info == NULL, mbedtls_ssl_tls1_3_derive_master_secret failed" ) );
-        return( MBEDTLS_ERR_SSL_INTERNAL_ERROR );
-    }
-
-    /* Determine hash size */
-    hash_size = mbedtls_hash_size_for_ciphersuite( suite_info );
-    if( hash_size == -1 )
-    {
-        MBEDTLS_SSL_DEBUG_MSG( 1, ( "mbedtls_hash_size_for_ciphersuite( ) failed" ) );
-        return( MBEDTLS_ERR_SSL_INTERNAL_ERROR );
-    }
-
-    /*
-     * Compute PSK for first stage of secret evolution.
-     */
-
-#if defined(MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED)
-    psk = ssl->conf->psk;
-    psk_len = ssl->conf->psk_len;
-#endif /* MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED */
-
-#if defined(MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED)
-    /* If the psk callback was called, use its result */
-    if( ( ssl->handshake->psk != NULL ) &&
-        ( ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_PSK ||
-          ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_ECDHE_PSK ) )
-    {
-        psk = ssl->handshake->psk;
-        psk_len = ssl->handshake->psk_len;
-    }
-#endif /* MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED */
-#if defined(MBEDTLS_KEY_EXCHANGE_ECDHE_ECDSA_ENABLED)
-    if( ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_ECDHE_ECDSA )
-    {
-
-        /* If we are not using a PSK-based ciphersuite then the
-         * psk identity is set to a 0 vector.
-         */
-
-        psk = mbedtls_calloc( hash_size,1 );
-        if( psk == NULL )
-        {
-            MBEDTLS_SSL_DEBUG_MSG( 1, ( "malloc for psk == NULL, mbedtls_ssl_tls1_3_derive_master_secret failed" ) );
-            return( MBEDTLS_ERR_SSL_INTERNAL_ERROR );
-        }
-        psk_allocated = 1;
-        memset( psk, 0x0, hash_size );
-        psk_len = hash_size;
-    }
-#endif /* MBEDTLS_KEY_EXCHANGE_ECDHE_ECDSA_ENABLED */
-
-    /*
-     * Compute ECDHE secret for second stage of secret evolution.
-     */
-
-#if defined(MBEDTLS_KEY_EXCHANGE_SOME_ECDHE_ENABLED)
-    if( ( ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_ECDHE_PSK ) ||
-        ( ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_ECDHE_ECDSA ) )
-    {
-        if( ( ret = mbedtls_ecdh_calc_secret( &ssl->handshake->ecdh_ctx[ssl->handshake->ecdh_ctx_selected],
-                                              &ECDHE_len,
-                                              ECDHE,
-                                              sizeof( ECDHE ),
-                                              ssl->conf->f_rng, ssl->conf->p_rng ) ) != 0 )
-        {
-            MBEDTLS_SSL_DEBUG_RET( 1, "mbedtls_ecdh_calc_secret", ret );
-            if( psk_allocated == 1 ) mbedtls_free( psk );
-            return( ret );
-        }
-
-        MBEDTLS_SSL_DEBUG_MPI( 3, "ECDHE:", &ssl->handshake->ecdh_ctx[ssl->handshake->ecdh_ctx_selected].z );
-
-    } else
-#endif /* MBEDTLS_KEY_EXCHANGE_SOME_ECDHE_ENABLED */
-#if defined(MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED)
-    if( ssl->session_negotiate->key_exchange == MBEDTLS_KEY_EXCHANGE_PSK )
-    {
-        memset( ECDHE, 0x0, hash_size );
-        MBEDTLS_SSL_DEBUG_BUF( 3, "ECDHE", ECDHE, hash_size );
-        ECDHE_len=hash_size;
-    } else
-#endif	/* MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED */
-    {
-        MBEDTLS_SSL_DEBUG_MSG( 1, ( "Unsupported key exchange -- mbedtls_ssl_tls1_3_derive_master_secret failed." ) );
-        if( psk_allocated == 1 ) mbedtls_free( psk );
-        return( MBEDTLS_ERR_SSL_BAD_HS_SERVER_HELLO );
-    }
-
-
-    /*
-     * Compute EarlySecret
-     */
-
-    ret = mbedtls_ssl_tls1_3_evolve_secret( ssl->handshake->ciphersuite_info->mac,
-                                            NULL, /* use 0 as old secret */
-                                            psk, psk_len,
-                                            ssl->handshake->early_secret );
-    if( ret != 0 )
-    {
-        MBEDTLS_SSL_DEBUG_RET( 1, "mbedtls_ssl_tls1_3_evolve_secret", ret );
-        if( psk_allocated == 1 )
-            mbedtls_free( psk );
-        return( ret );
-    }
-
-    MBEDTLS_SSL_DEBUG_BUF( 4, "Early secret", ssl->handshake->early_secret, hash_size );
-
-    /*
-     * Compute HandshakeSecret
-     */
-
-    ret = mbedtls_ssl_tls1_3_evolve_secret(
-                              ssl->handshake->ciphersuite_info->mac,
-                              ssl->handshake->early_secret,
-                              ECDHE, ECDHE_len,
-                              ssl->handshake->handshake_secret );
-    if( ret != 0 )
-    {
-        MBEDTLS_SSL_DEBUG_RET( 1, "mbedtls_hkdf_extract( ) with early_secret", ret );
-        if( psk_allocated == 1 ) mbedtls_free( psk );
-        return( ret );
-    }
-
-    MBEDTLS_SSL_DEBUG_BUF( 4, "Handshake secret", ssl->handshake->handshake_secret, hash_size );
-
-    /*
-     * Compute MasterSecret
-     */
-
-    ret = mbedtls_ssl_tls1_3_evolve_secret(
-                              ssl->handshake->ciphersuite_info->mac,
-                              ssl->handshake->handshake_secret,
-                              NULL, 0,
-                              ssl->handshake->master_secret );
-    if( ret != 0 )
-    {
-        MBEDTLS_SSL_DEBUG_RET( 1, "mbedtls_hkdf_extract( ) with early_secret", ret );
-        if( psk_allocated == 1 ) mbedtls_free( psk );
-        return( ret );
-    }
-
-    MBEDTLS_SSL_DEBUG_BUF( 4, "Master secret", ssl->handshake->master_secret, hash_size );
-
-    if( psk_allocated == 1 ) mbedtls_free( psk );
-
-    /*
-     * Export client early traffic secret
-     */
-#if defined(MBEDTLS_SSL_EXPORT_KEYS)
-    if( ssl->conf->f_export_secret != NULL )
-    {
-        ssl->conf->f_export_secret( ssl->conf->p_export_secret,
-                ssl->handshake->randbytes,
-                MBEDTLS_SSL_TLS1_3_CLIENT_EARLY_TRAFFIC_SECRET,
-                ssl->handshake->early_secret, hash_size );
-    }
-#endif /* MBEDTLS_SSL_EXPORT_KEYS */
-
-
-    return( 0 );
-}
-
-
-
 /*
  *
  * STATE HANDLING: CertificateVerify
