+++ library/ssl_tls13_keys.h
@@ -321,4 +321,6 @@ int mbedtls_ssl_create_binder( mbedtls_ssl_context *ssl,
                                unsigned char *result );
 #endif /* MBEDTLS_KEY_EXCHANGE_SOME_PSK_ENABLED */
 
+int mbedtls_ssl_tls1_3_derive_master_secret( mbedtls_ssl_context* ssl );
+
 #endif /* MBEDTLS_SSL_TLS1_3_KEYS_H */
